Units and Workout tests now in two files.  Changed Workout db test name
